--- multifile/paster/__init__.py
@@ -1,2 +1,0 @@
-from .incremental import *
-from .naive import *
