--- requirements/dev.txt
@@ -1,1 +1,0 @@
-pytest==5.3.4
