--- scripts/no_install/paste_vcf.py
@@ -1,8 +1,0 @@
-from pathlib import Path
-import sys
-
-sys.path.append(str(Path(__file__).absolute().parent.parent.parent))
-
-from multifile.paster.vcf.__main__ import main
-
-main()
