--- tests/data/ref_batch_merged.vcf
@@ -1,20 +1,0 @@
-##fileformat=VCFv4.2##fileformat=VCFv4.2##fileformat=VCFv4.2##fileformat=VCFv4.2
-##source=minos, version 0.9.1##source=minos, version 0.9.1##source=minos, version 0.9.1##source=minos, version 0.9.1
-##fileDate=2020-01-21##fileDate=2020-01-21##fileDate=2020-01-21##fileDate=2020-01-21
-##FORMAT=<ID=COV,Number=R,Type=Integer,Description="Number of reads on ref and alt alleles">##FORMAT=<ID=COV,Number=R,Type=Integer,Description="Number of reads on ref and alt alleles">##FORMAT=<ID=COV,Number=R,Type=Integer,Description="Number of reads on ref and alt alleles">##FORMAT=<ID=COV,Number=R,Type=Integer,Description="Number of reads on ref and alt alleles">
-##FORMAT=<ID=GT,Number=1,Type=String,Description="Genotype">##FORMAT=<ID=GT,Number=1,Type=String,Description="Genotype">##FORMAT=<ID=GT,Number=1,Type=String,Description="Genotype">##FORMAT=<ID=GT,Number=1,Type=String,Description="Genotype">
-##FORMAT=<ID=DP,Number=1,Type=Integer,Description="total read depth from gramtools">##FORMAT=<ID=DP,Number=1,Type=Integer,Description="total read depth from gramtools">##FORMAT=<ID=DP,Number=1,Type=Integer,Description="total read depth from gramtools">##FORMAT=<ID=DP,Number=1,Type=Integer,Description="total read depth from gramtools">
-##FORMAT=<ID=GT_CONF,Number=1,Type=Float,Description="Genotype confidence. Difference in log likelihood of most likely and next most likely genotype">##FORMAT=<ID=GT_CONF,Number=1,Type=Float,Description="Genotype confidence. Difference in log likelihood of most likely and next most likely genotype">##FORMAT=<ID=GT_CONF,Number=1,Type=Float,Description="Genotype confidence. Difference in log likelihood of most likely and next most likely genotype">##FORMAT=<ID=GT_CONF,Number=1,Type=Float,Description="Genotype confidence. Difference in log likelihood of most likely and next most likely genotype">
-##FILTER=<ID=MIN_GCP,Description="Minimum GT_CONF_PERCENTILE of 5.0">##FILTER=<ID=MIN_GCP,Description="Minimum GT_CONF_PERCENTILE of 5.0">##FILTER=<ID=MIN_GCP,Description="Minimum GT_CONF_PERCENTILE of 5.0">##FILTER=<ID=MIN_GCP,Description="Minimum GT_CONF_PERCENTILE of 5.0">
-##FILTER=<ID=MIN_DP,Description="Minimum DP of 5">##FILTER=<ID=MIN_DP,Description="Minimum DP of 5">##FILTER=<ID=MIN_DP,Description="Minimum DP of 5">##FILTER=<ID=MIN_DP,Description="Minimum DP of 5">
-##FORMAT=<ID=GT_CONF_PERCENTILE,Number=1,Type=Float,Description="Percentile of GT_CONF">##FORMAT=<ID=GT_CONF_PERCENTILE,Number=1,Type=Float,Description="Percentile of GT_CONF">##FORMAT=<ID=GT_CONF_PERCENTILE,Number=1,Type=Float,Description="Percentile of GT_CONF">##FORMAT=<ID=GT_CONF_PERCENTILE,Number=1,Type=Float,Description="Percentile of GT_CONF">
-##minos_max_read_length=150##minos_max_read_length=151##minos_max_read_length=151##minos_max_read_length=151
-#CHROM	POS	ID	REF	ALT	QUAL	FILTER	INFO	FORMAT	1#CHROM	POS	ID	REF	ALT	QUAL	FILTER	INFO	FORMAT	2#CHROM	POS	ID	REF	ALT	QUAL	FILTER	INFO	FORMAT	2#CHROM	POS	ID	REF	ALT	QUAL	FILTER	INFO	FORMAT	4
-NC_000962.3	11	.	A	C	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:6:6,0:73.54:0.74NC_000962.3	11	.	A	C	.	MIN_DP;MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:3:3,0:36.98:0.01NC_000962.3	11	.	A	C	.	MIN_DP;MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:3:3,0:36.98:0.01NC_000962.3	11	.	A	C	.	MIN_DP;MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:2:2,0:23.4:0.0
-NC_000962.3	19	.	T	A	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:9:9,0:198.51:1.1NC_000962.3	19	.	T	A	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:12:12,0:245.33:2.04NC_000962.3	19	.	T	A	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:12:12,0:245.33:2.04NC_000962.3	19	.	T	A	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:9:9,0:168.6:4.08
-NC_000962.3	22	.	G	C	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:11:11,0:220.52:1.1NC_000962.3	22	.	G	C	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:13:13,0:255.69:2.09NC_000962.3	22	.	G	C	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:13:13,0:255.69:2.09NC_000962.3	22	.	G	C	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:10:10,0:178.35:4.08
-NC_000962.3	29	.	C	G	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:15:15,0:263.53:1.1NC_000962.3	29	.	C	G	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:17:17,0:296.43:2.64NC_000962.3	29	.	C	G	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:17:17,0:296.43:2.64NC_000962.3	29	.	C	G	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:22:22,0:289.57:4.08
-NC_000962.3	31	.	A	G	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:16:16,0:274.12:1.1NC_000962.3	31	.	A	G	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:18:18,0:306.46:2.79NC_000962.3	31	.	A	G	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:18:18,0:306.46:2.79NC_000962.3	31	.	A	G	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:23:23,0:298.48:4.08
-NC_000962.3	40	.	A	T	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:23:23,0:346.66:1.11NC_000962.3	40	.	A	T	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:25:25,0:375.31:4.92NC_000962.3	40	.	A	T	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:25:25,0:375.31:4.92NC_000962.3	40	.	A	T	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:25:25,0:316.18:4.08
-NC_000962.3	61	.	A	G	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:31:31,0:427.0:1.19NC_000962.3	61	.	A	G	.	PASS	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:30:30,0:423.27:7.75NC_000962.3	61	.	A	G	.	PASS	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:30:30,0:423.27:7.75NC_000962.3	61	.	A	G	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:33:33,0:385.52:4.08
-NC_000962.3	64	.	G	C	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:33:33,0:446.75:1.34NC_000962.3	64	.	G	C	.	PASS	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:30:30,0:423.27:7.75NC_000962.3	64	.	G	C	.	PASS	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:30:30,0:423.27:7.75NC_000962.3	64	.	G	C	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:33:33,0:385.52:4.08
