--- tests/data/ref_fixed_cols.vcf
@@ -1,20 +1,0 @@
-##fileformat=VCFv4.2
-##source=minos, version 0.9.1
-##fileDate=2020-01-21
-##FORMAT=<ID=COV,Number=R,Type=Integer,Description="Number of reads on ref and alt alleles">
-##FORMAT=<ID=GT,Number=1,Type=String,Description="Genotype">
-##FORMAT=<ID=DP,Number=1,Type=Integer,Description="total read depth from gramtools">
-##FORMAT=<ID=GT_CONF,Number=1,Type=Float,Description="Genotype confidence. Difference in log likelihood of most likely and next most likely genotype">
-##FILTER=<ID=MIN_GCP,Description="Minimum GT_CONF_PERCENTILE of 5.0">
-##FILTER=<ID=MIN_DP,Description="Minimum DP of 5">
-##FORMAT=<ID=GT_CONF_PERCENTILE,Number=1,Type=Float,Description="Percentile of GT_CONF">
-##minos_max_read_length=150
-#CHROM	POS	ID	REF	ALT	QUAL	FILTER	INFO	FORMAT	1	2
-NC_000962.3	11	.	A	C	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:6:6,0:73.54:0.74	0/0:3:3,0:36.98:0.01
-NC_000962.3	19	.	T	A	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:9:9,0:198.51:1.1	0/0:12:12,0:245.33:2.04
-NC_000962.3	22	.	G	C	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:11:11,0:220.52:1.1	0/0:13:13,0:255.69:2.09
-NC_000962.3	29	.	C	G	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:15:15,0:263.53:1.1	0/0:17:17,0:296.43:2.64
-NC_000962.3	31	.	A	G	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:16:16,0:274.12:1.1	0/0:18:18,0:306.46:2.79
-NC_000962.3	40	.	A	T	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:23:23,0:346.66:1.11	0/0:25:25,0:375.31:4.92
-NC_000962.3	61	.	A	G	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:31:31,0:427.0:1.19	0/0:30:30,0:423.27:7.75
-NC_000962.3	64	.	G	C	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:33:33,0:446.75:1.34	0/0:30:30,0:423.27:7.75
