--- tests/data/ref_fixed_cols_incremental.vcf
@@ -1,20 +1,0 @@
-##fileformat=VCFv4.2
-##source=minos, version 0.9.1
-##fileDate=2020-01-21
-##FORMAT=<ID=COV,Number=R,Type=Integer,Description="Number of reads on ref and alt alleles">
-##FORMAT=<ID=GT,Number=1,Type=String,Description="Genotype">
-##FORMAT=<ID=DP,Number=1,Type=Integer,Description="total read depth from gramtools">
-##FORMAT=<ID=GT_CONF,Number=1,Type=Float,Description="Genotype confidence. Difference in log likelihood of most likely and next most likely genotype">
-##FILTER=<ID=MIN_GCP,Description="Minimum GT_CONF_PERCENTILE of 5.0">
-##FILTER=<ID=MIN_DP,Description="Minimum DP of 5">
-##FORMAT=<ID=GT_CONF_PERCENTILE,Number=1,Type=Float,Description="Percentile of GT_CONF">
-##minos_max_read_length=151
-#CHROM	POS	ID	REF	ALT	QUAL	FILTER	INFO	FORMAT	5	6	1	2	2	4
-NC_000962.3	11	.	A	C	.	.	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	./.:0:0,0:0.0:0.0	0/0:3:3,0:32.98:0.01	0/0:6:6,0:73.54:0.74	0/0:3:3,0:36.98:0.01	0/0:3:3,0:36.98:0.01	0/0:2:2,0:23.4:0.0
-NC_000962.3	19	.	T	A	.	.	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	./.:0:0,0:0.0:0.0	0/0:7:7,0:117.39:2.39	0/0:9:9,0:198.51:1.1	0/0:12:12,0:245.33:2.04	0/0:12:12,0:245.33:2.04	0/0:9:9,0:168.6:4.08
-NC_000962.3	22	.	G	C	.	.	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	./.:0:0,0:0.0:0.0	0/0:8:8,0:126.9:2.39	0/0:11:11,0:220.52:1.1	0/0:13:13,0:255.69:2.09	0/0:13:13,0:255.69:2.09	0/0:10:10,0:178.35:4.08
-NC_000962.3	29	.	C	G	.	MIN_DP;MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:2:2,0:28.49:1.11	0/0:10:10,0:145.58:2.39	0/0:15:15,0:263.53:1.1	0/0:17:17,0:296.43:2.64	0/0:17:17,0:296.43:2.64	0/0:22:22,0:289.57:4.08
-NC_000962.3	31	.	A	G	.	MIN_DP;MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:2:2,0:28.49:1.11	0/0:12:12,0:163.87:2.39	0/0:16:16,0:274.12:1.1	0/0:18:18,0:306.46:2.79	0/0:18:18,0:306.46:2.79	0/0:23:23,0:298.48:4.08
-NC_000962.3	40	.	A	T	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:7:7,0:69.94:1.59	0/0:16:16,0:199.54:2.4	0/0:23:23,0:346.66:1.11	0/0:25:25,0:375.31:4.92	0/0:25:25,0:375.31:4.92	0/0:25:25,0:316.18:4.08
-NC_000962.3	61	.	A	G	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:10:10,0:92.94:2.47	0/0:23:23,0:259.73:2.47	0/0:31:31,0:427.0:1.19	0/0:30:30,0:423.27:7.75	0/0:30:30,0:423.27:7.75	0/0:33:33,0:385.52:4.08
-NC_000962.3	64	.	G	C	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:10:10,0:92.94:2.47	0/0:23:23,0:259.73:2.47	0/0:33:33,0:446.75:1.34	0/0:30:30,0:423.27:7.75	0/0:30:30,0:423.27:7.75	0/0:33:33,0:385.52:4.08
