--- tests/data/ref_invalid.vcf
@@ -1,19 +1,0 @@
-##source=minos, version 0.9.1
-##fileDate=2020-01-21
-##FORMAT=<ID=COV,Number=R,Type=Integer,Description="Number of reads on ref and alt alleles">
-##FORMAT=<ID=GT,Number=1,Type=String,Description="Genotype">
-##FORMAT=<ID=DP,Number=1,Type=Integer,Description="total read depth from gramtools">
-##FORMAT=<ID=GT_CONF,Number=1,Type=Float,Description="Genotype confidence. Difference in log likelihood of most likely and next most likely genotype">
-##FILTER=<ID=MIN_GCP,Description="Minimum GT_CONF_PERCENTILE of 5.0">
-##FILTER=<ID=MIN_DP,Description="Minimum DP of 5">
-##FORMAT=<ID=GT_CONF_PERCENTILE,Number=1,Type=Float,Description="Percentile of GT_CONF">
-##minos_max_read_length=150
-#CHROM	POS	ID	REF	ALT	QUAL	FILTER	INFO	FORMAT	invalid_sample_name
-NC_000962.3	11	.	A	C	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:6:6,0:73.54:0.74
-NC_000962.3	19	.	T	A	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:9:9,0:198.51:1.1
-NC_000962.3	22	.	G	C	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:11:11,0:220.52:1.1
-NC_000962.3	29	.	C	G	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:15:15,0:263.53:1.1
-NC_000962.3	31	.	A	G	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:16:16,0:274.12:1.1
-NC_000962.3	40	.	A	T	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:23:23,0:346.66:1.11
-NC_000962.3	61	.	A	G	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:31:31,0:427.0:1.19
-NC_000962.3	64	.	G	C	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:33:33,0:446.75:1.34
