--- tests/data/ref_skipped_headers.vcf
@@ -1,9 +1,0 @@
-#CHROM	POS	ID	REF	ALT	QUAL	FILTER	INFO	FORMAT	1#CHROM	POS	ID	REF	ALT	QUAL	FILTER	INFO	FORMAT	2
-NC_000962.3	11	.	A	C	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:6:6,0:73.54:0.74NC_000962.3	11	.	A	C	.	MIN_DP;MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:3:3,0:36.98:0.01
-NC_000962.3	19	.	T	A	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:9:9,0:198.51:1.1NC_000962.3	19	.	T	A	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:12:12,0:245.33:2.04
-NC_000962.3	22	.	G	C	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:11:11,0:220.52:1.1NC_000962.3	22	.	G	C	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:13:13,0:255.69:2.09
-NC_000962.3	29	.	C	G	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:15:15,0:263.53:1.1NC_000962.3	29	.	C	G	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:17:17,0:296.43:2.64
-NC_000962.3	31	.	A	G	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:16:16,0:274.12:1.1NC_000962.3	31	.	A	G	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:18:18,0:306.46:2.79
-NC_000962.3	40	.	A	T	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:23:23,0:346.66:1.11NC_000962.3	40	.	A	T	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:25:25,0:375.31:4.92
-NC_000962.3	61	.	A	G	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:31:31,0:427.0:1.19NC_000962.3	61	.	A	G	.	PASS	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:30:30,0:423.27:7.75
-NC_000962.3	64	.	G	C	.	MIN_GCP	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:33:33,0:446.75:1.34NC_000962.3	64	.	G	C	.	PASS	.	GT:DP:COV:GT_CONF:GT_CONF_PERCENTILE	0/0:30:30,0:423.27:7.75
