--- tests/data/ref_stripped_cols.vcf
@@ -1,9 +1,0 @@
-12
-0/0:6:6,0:73.54:0.740/0:3:3,0:36.98:0.01
-0/0:9:9,0:198.51:1.10/0:12:12,0:245.33:2.04
-0/0:11:11,0:220.52:1.10/0:13:13,0:255.69:2.09
-0/0:15:15,0:263.53:1.10/0:17:17,0:296.43:2.64
-0/0:16:16,0:274.12:1.10/0:18:18,0:306.46:2.79
-0/0:23:23,0:346.66:1.110/0:25:25,0:375.31:4.92
-0/0:31:31,0:427.0:1.190/0:30:30,0:423.27:7.75
-0/0:33:33,0:446.75:1.340/0:30:30,0:423.27:7.75
